# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "spjall-api",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "deviceToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "The device token from registerDevice (decision 0014)"
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "A stable code: unauthorized, invalid_request, group_mismatch, not_a_member, not_found, conversation_exists, epoch_conflict, upgrade_required, sign_in_failed, invite_required, device_key_taken, kenni_unavailable, rate_limited, blocked, conflict, too_large, not_this_device, no_key_packages, client_too_old, claim_names_departed",
            "example": "unauthorized"
          },
          "minVersion": {
            "type": "string",
            "description": "With client_too_old: the lowest version this platform accepts (decision 0030)",
            "example": "0.2.0"
          }
        },
        "required": [
          "error"
        ]
      },
      "RegisteredDevice": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,128}$",
            "example": "c_9f3a1b"
          },
          "deviceId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,128}$",
            "example": "c_9f3a1b"
          },
          "token": {
            "type": "string",
            "description": "The device token, shown once; send it as Bearer"
          }
        },
        "required": [
          "accountId",
          "deviceId",
          "token"
        ]
      },
      "RegisterDevice": {
        "type": "object",
        "properties": {
          "kenniCode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048,
            "description": "The Kenni authorization code"
          },
          "codeVerifier": {
            "type": "string",
            "pattern": "^[A-Za-z0-9._~-]{43,128}$",
            "description": "The PKCE verifier for that code"
          },
          "redirectUri": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512,
            "description": "The redirect the code was issued for: SignInConfig.redirectUri"
          },
          "nonce": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{22,128}$",
            "description": "The nonce the app put in the authorization request"
          },
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "ios"
            ]
          },
          "deviceKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 44,
            "pattern": "^(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{2}==|[A-Za-z0-9+/]{3}=)?$",
            "description": "The device's Ed25519 public key, also its MLS credential key"
          },
          "inviteToken": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256,
            "description": "Required when the Kenni sign-in has no account yet (decision 0009)"
          }
        },
        "required": [
          "kenniCode",
          "codeVerifier",
          "redirectUri",
          "nonce",
          "platform",
          "deviceKey"
        ]
      },
      "Me": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,128}$",
            "example": "c_9f3a1b"
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "The registry name, if Kenni gave one"
          },
          "verified": {
            "type": "boolean",
            "description": "Kenni vouched for the name (decision 0009)"
          },
          "devices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "deviceId": {
                  "type": "string",
                  "pattern": "^[A-Za-z0-9_-]{1,128}$",
                  "example": "c_9f3a1b"
                },
                "platform": {
                  "type": "string",
                  "enum": [
                    "android",
                    "ios"
                  ]
                },
                "createdAt": {
                  "type": "integer",
                  "description": "Milliseconds since the epoch"
                },
                "current": {
                  "type": "boolean",
                  "description": "The device that asked"
                }
              },
              "required": [
                "deviceId",
                "platform",
                "createdAt",
                "current"
              ]
            }
          }
        },
        "required": [
          "accountId",
          "name",
          "verified",
          "devices"
        ]
      },
      "NewInvite": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{16,128}$",
            "example": "q1w2e3r4t5y6u7i8o9p0aAsSdDfFgGhHjJkKlLzZxXc"
          },
          "link": {
            "type": "string",
            "description": "The link to share; shown once, since the server keeps only its hash",
            "example": "https://spjall.samtak.is/l/q1w2e3r4t5y6u7i8o9p0aAsSdDfFgGhHjJkKlLzZxXc"
          }
        },
        "required": [
          "token",
          "link"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/devices": {
      "post": {
        "operationId": "registerDevice",
        "tags": [
          "devices"
        ],
        "summary": "Sign in with Kenni and register this device",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDevice"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The device is registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredDevice"
                }
              }
            }
          },
          "400": {
            "description": "The request does not match the schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "sign_in_failed: Kenni refused the code or its ID token did not verify; invite_required: no account for this person and no live invite",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "device_key_taken: another device registered this device key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "426": {
            "description": "client_too_old: Spjall-Client names a version below the floor; minVersion says which",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: too many requests in the last minute; try again later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "kenni_unavailable: Kenni could not be reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "getMe",
        "tags": [
          "devices"
        ],
        "summary": "This account: its name, its mark and its active devices",
        "security": [
          {
            "deviceToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "The account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "401": {
            "description": "No device token, or one that is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "426": {
            "description": "client_too_old: Spjall-Client names a version below the floor; minVersion says which",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteAccount",
        "tags": [
          "devices"
        ],
        "summary": "Delete this account, its devices, inbox and media (decision 0014)",
        "description": "Every token is revoked, each conversation drops the account from its roster, the inbox is deleted and the sockets close with 4401, then the account's rows go, with its invites (decision 0019).",
        "security": [
          {
            "deviceToken": []
          }
        ],
        "responses": {
          "204": {
            "description": "The account is deleted"
          },
          "401": {
            "description": "No device token, or one that is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "426": {
            "description": "client_too_old: Spjall-Client names a version below the floor; minVersion says which",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/invite": {
      "post": {
        "operationId": "rotateInvite",
        "tags": [
          "invites"
        ],
        "summary": "Make this account's invite link, revoking the one before",
        "security": [
          {
            "deviceToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "The new link; the old one no longer lets anyone in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewInvite"
                }
              }
            }
          },
          "401": {
            "description": "No device token, or one that is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "426": {
            "description": "client_too_old: Spjall-Client names a version below the floor; minVersion says which",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "revokeInvite",
        "tags": [
          "invites"
        ],
        "summary": "Revoke this account's invite link and leave it without one",
        "security": [
          {
            "deviceToken": []
          }
        ],
        "responses": {
          "204": {
            "description": "The account has no live invite"
          },
          "401": {
            "description": "No device token, or one that is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "426": {
            "description": "client_too_old: Spjall-Client names a version below the floor; minVersion says which",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/invites/{token}": {
      "get": {
        "operationId": "resolveInvite",
        "tags": [
          "invites"
        ],
        "summary": "Who sent an invite link; needs no device token, the token is the capability",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{16,128}$",
              "example": "q1w2e3r4t5y6u7i8o9p0aAsSdDfFgGhHjJkKlLzZxXc"
            },
            "required": true,
            "name": "token",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The invite is live",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedInvite"
                }
              }
            }
          },
          "400": {
            "description": "The request does not match the schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "not_found: no live invite with this token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "426": {
            "description": "client_too_old: Spjall-Client names a version below the floor; minVersion says which",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: too many requests in the last minute; try again later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/devices/{deviceId}": {
      "delete": {
        "operationId": "revokeDevice",
        "tags": [
          "devices"
        ],
        "summary": "Revoke a device of this account and its token",
        "description": "The token fails at once, the device's KeyPackages are deleted, and its socket closes with 4401. A device may revoke itself; that is signing out (decision 0019).",
        "security": [
          {
            "deviceToken": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$",
              "example": "c_9f3a1b"
            },
            "required": true,
            "name": "deviceId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "The device is revoked"
          },
          "400": {
            "description": "The request does not match the schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "No device token, or one that is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "not_found: this account has no such active device",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "426": {
            "description": "client_too_old: Spjall-Client names a version below the floor; minVersion says which",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}